+++ lib/vagrant/util/platform.rb
@@ -44,9 +44,14 @@ module Vagrant
           !bit64?
         end
 
+        # Returns a boolean noting whether the terminal supports color.
+        # output.
         def terminal_supports_colors?
-           !$stdout.tty? || ARGV.include?("--no-color") || 
-             (Vagrant::Util::Platform.windows? ? ENV['ansicon'] : false)
+          if windows?
+            return ENV.has_key?("ANSICON")
+          end
+
+          true
         end
 
         def tar_file_options